+++ src/slurmd/slurmstepd/slurmstepd_job.h
@@ -137,8 +137,8 @@ typedef struct {
 	uint16_t      *task_cnts;  /* Number of tasks on each node in job   */
 	uint32_t       cpus_per_task;	/* number of cpus desired per task  */
 	uint32_t       debug;  /* debug level for job slurmd                */
-	uint32_t       job_mem;  /* MB of memory reserved for the job       */
-	uint32_t       step_mem; /* MB of memory reserved for the step      */
+	uint64_t       job_mem;  /* MB of memory reserved for the job       */
+	uint64_t       step_mem; /* MB of memory reserved for the step      */
 	uint16_t       cpus;   /* number of cpus to use for this job        */
 	uint32_t       argc;   /* number of commandline arguments           */
 	char         **env;    /* job environment                           */
